seller views separated
