+++ ecommerce_project/myapp/views/__init__.py
@@ -5,3 +5,4 @@ from .product_views import *
 from .review_views import *
 from .order_views import *
 from .buyer_views import *
+from .seller_views import *
